requirements.txt
Flask==3.0.0
Pillow==10.1.0
Werkzeug==3.0.1
matplotlib==3.8.2
numpy==1.26.2
gunicorn==21.2.0
